Add files via upload
This file Project_java.zip has the client server set up that allows you to enter a port number to sign into a server. It is basically a sign up pages that gets all the information from the user. It runs a password validation check to make sure it is the same password before submitting the information to the server.