Sign up page in java
This is the updated version for the sign up page. I am still new to using GitHub and don't know how to remove my previous upload. This file is similar to the last one, but it will allow the client to connect to the server and this will send the user's information to the server side for validation.